+++ src/main/kotlin/org/stjude/sprocket/lsp/SprocketLanguageServerFactory.kt
@@ -4,6 +4,7 @@ import com.intellij.openapi.project.Project
 import com.redhat.devtools.lsp4ij.LanguageServerFactory
 import com.redhat.devtools.lsp4ij.client.LanguageClientImpl
 import com.redhat.devtools.lsp4ij.server.StreamConnectionProvider
+import org.stjude.sprocket.client.WdlLanguageClient
 
 class SprocketLanguageServerFactory : LanguageServerFactory {
 
@@ -11,5 +12,5 @@ class SprocketLanguageServerFactory : LanguageServerFactory {
         SprocketStreamConnectionProvider(project)
 
     override fun createLanguageClient(project: Project): LanguageClientImpl =
-        LanguageClientImpl(project)
+        WdlLanguageClient(project)
 }